moving variable specific code to own class
