+++ .vscode/launch.json
@@ -20,6 +20,23 @@
             "outFiles": ["${workspaceFolder}/out/**/*.js"]
         },
         {
+            "name": "Test (Coverage)",
+            "type": "node",
+            "request": "launch",
+            "protocol": "inspector",
+            "program": "${workspaceFolder}/node_modules/nyc/bin/nyc",
+            "args": [
+                "mocha",
+                "--require", "source-map-support/register",  "--exit", "--recursive", "-u", "tdd", "./out/test/"
+            ],
+            "cwd": "${workspaceFolder}",
+            "smartStep": true,
+            "skipFiles": [
+                "<node_internals>/**/*.js"
+            ],
+            "outFiles": ["${workspaceFolder}/out/**/*.js"]
+        },
+        {
             "name": "Grepped Tests",
             "type": "node",
             "request": "launch",
